move warehouse_update to its own module so it can be further split up
